+++ src/main/java/ua/lviv/iot/bank/services/DepositService.java
@@ -5,15 +5,31 @@ import org.springframework.web.bind.annotation.*;
 import ua.lviv.iot.bank.models.Deposit;
 import ua.lviv.iot.bank.repositories.DepositRepository;
 
+import java.util.LinkedList;
+import java.util.List;
+
 @RestController
 public class DepositService {
 
     @Autowired
     private DepositRepository repository;
 
+    @RequestMapping(value = "/deposit/", method = RequestMethod.GET)
+    public List<Deposit> getAllDeposits() {
+        List<Deposit> deposits = new LinkedList<>();
+        repository.findAll().forEach(deposits::add);
+        if(deposits.isEmpty()) {
+            return null;
+        }
+        return deposits;
+    }
+
     @RequestMapping(value = "/deposit/{id}", method = RequestMethod.GET)
     public Deposit getDeposit(@PathVariable Integer id) {
-        return repository.findById(id).get();
+        if (repository.findById(id).isPresent()) {
+            return repository.findById(id).get();
+        }
+        return null;
     }
 
     @RequestMapping(value = "/deposit", method = RequestMethod.POST)
@@ -28,6 +44,8 @@ public class DepositService {
 
     @RequestMapping(value = "/deposit/{id}", method = RequestMethod.DELETE)
     public void deleteDeposit(@PathVariable Integer id) {
-        repository.deleteById(id);
+        if (repository.findById(id).isPresent()) {
+            repository.deleteById(id);
+        }
     }
 }
