+++ src/main/java/ua/lviv/iot/bank/services/RemittanceService.java
@@ -5,15 +5,31 @@ import org.springframework.web.bind.annotation.*;
 import ua.lviv.iot.bank.models.Remittance;
 import ua.lviv.iot.bank.repositories.RemittanceRepository;
 
+import java.util.LinkedList;
+import java.util.List;
+
 @RestController
 public class RemittanceService {
 
     @Autowired
     private RemittanceRepository repository;
 
+    @RequestMapping(value = "/deposit/", method = RequestMethod.GET)
+    public List<Remittance> getAllDeposits() {
+        List<Remittance> remittances = new LinkedList<>();
+        repository.findAll().forEach(remittances::add);
+        if(remittances.isEmpty()) {
+            return null;
+        }
+        return remittances;
+    }
+
     @RequestMapping(value = "/remittance/{id}", method = RequestMethod.GET)
     public Remittance getRemittance(@PathVariable Integer id) {
-        return repository.findById(id).get();
+        if (repository.findById(id).isPresent()) {
+            return repository.findById(id).get();
+        }
+        return null;
     }
 
     @RequestMapping(value = "/remittance", method = RequestMethod.POST)
@@ -28,6 +44,8 @@ public class RemittanceService {
 
     @RequestMapping(value = "/remittance/{id}", method = RequestMethod.DELETE)
     public void deleteRemittance(@PathVariable Integer id) {
-        repository.deleteById(id);
+        if (repository.findById(id).isPresent()) {
+            repository.deleteById(id);
+        }
     }
 }
